katharanp: use bridge set in option
Instead of creating (and deleting) a bridge automatically, the user must
now create the bridge, and set its name when creating the network with
-o bridge=bridgename.

By this way, the user can now a full control on the bridge, and he can
bind external interface directly to the bridge.

Note: --cap_add NET_ADMIN should be configured for a given container
(docker run), in order to allow the network configuration.

Signed-off-by: Julien Meunier <julien.meunier@nokia.com>